Refactor project.
