+++ src/api/tests/mod.rs
@@ -1,0 +1,1 @@
+mod gen_draw_fn;
