+++ src/rasterization/tests/mod.rs
@@ -1,0 +1,1 @@
+//! Test module for the rasterization module
