+++ src/types/tests/mod.rs
@@ -1,0 +1,2 @@
+mod line;
+mod point;
